Fix Modbus Address divide-by-10 issue
The Modbus address was showing as 1 instead of the expected default value of 10.
This was because all number values (except time values) were being divided by 10,
which is correct for voltage/current but incorrect for Modbus addresses.

Added an 'is_raw_value' flag to the ModbusAddressNumber class and updated
the base class methods to check for this flag when deciding whether to divide
by 10. Now the Modbus address will be displayed as a raw integer value,
matching the device's actual setting (default = 10).